Categorize polygon by Province/State if applicable

Categorize polygon by Province/State if applicable

Usage

assign_prov_state(
  strata_map,
  min_overlap = 0.75,
  plot = FALSE,
  keep_spatial = TRUE
)

Arguments

strata_map

sf data frame. Strata polygons to be categorized.

min_overlap

Numeric. The minimum proportion of overlap between a stratum polygon and a Province or State. Below this will raise warnings.

plot

Logical. Whether to plot how polygons were assigned to Provinces or States

keep_spatial

Logical. Whether the output should be a spatial data frame or not.

Value

(Spatial) data frame with strata assigned to Province/State

Examples

# Demonstration of why we can't divide BCR by Provinces and States!
map <- load_map("bcr")
assign_prov_state(map, plot = TRUE)
#> Error in stopifnot(!inherits(x, "sf"), !missing(sf_column_name), !missing(agr)): Problem while computing `prov_state = if_else(.data$prov_state == "NF",
#> "NL", .data$prov_state)`.
#> Caused by error in `if_else()`:
#> ! could not find function "if_else"

# Use custom stratification, using sf map object
# e.g. with WBPHS stratum boundaries 2019
# available: https://ecos.fws.gov/ServCat/Reference/Profile/142628

if (FALSE) {
map <- sf::read_sf("../WBPHS_Stratum_Boundaries_2019") %>%
  rename(strata_name = STRAT) # expects this column
s <- assign_prov_state(map, plot = TRUE)
# Some don't divide nicely, we could try a different min_overlap

s <- assign_prov_state(map, min_overlap = 0.6, plot = TRUE)
}