Generate regional trends

Generates trends for continent and strata and optionally for countries, states/provinces, or BCRs from analyses run on the stratifications that support these composite regions. Calculates the geometric mean annual changes in population size for composite regions.

Usage

generate_trends(
  indices,
  min_year = NULL,
  max_year = NULL,
  quantiles = c(0.025, 0.05, 0.25, 0.75, 0.95, 0.975),
  slope = FALSE,
  prob_decrease = NULL,
  prob_increase = NULL
)

Arguments

indices

List. Indices generated by generate_indices()

min_year

Numeric. Minimum year to use. Default (NULL) uses first year in data.

max_year

Numeric. Maximum year to use. Default (NULL) uses first year in data.

quantiles

Numeric vector. Quantiles to be sampled from the posterior distribution. Defaults to c(0.025, 0.05, 0.25, 0.5, 0.75, 0.95, 0.975)

slope

Logical. Whether to calculate an alternative trend metric, the slope of a log-linear regression through the annual indices. Default FALSE.

prob_decrease

Numeric vector. Percent-decrease values for which to optionally calculate the posterior probabilities (see Details). Default is NULL (do not calculate).

prob_increase

Numeric vector. Percent-increase values for which to optionally calculate the posterior probabilities (see Details). Default is NULL (do not calculate).

Value

Data frame with one row for each region included in the starting indices object. Contains start_year, end_year, region, region-alt, region_type, strata_included, strata_excluded

Start_year

first year of the trend

End_year

last year of the trend

Region

short name of the region

Region_alt

Long name for region

Region_type

Type of region including continental, national,Province_State,BCR, bcr_by_national, or stratum

Strata_included

Strata included in the trend and annual index calculations

Strata_excluded

Strata potentially excluded from the trend and annual index calculations because they have no observations of the species in the first part of the time series

Trend

Estimated mean annual percent change over the trend time-period (i.e., Start_year - End_year), according to an endpoint comparison of annual index in Start_year and the annual index in End_year

Trend_Q_quantiles

quantiles of the posterior distribution of Trend estimates, matching levels included in the quantiles argument

Percent_Change

Estimated total percent change over the trend time-period

Percent_Change_Q_quantiles

quantiles of the posterior distribution of Percent Change estimates, matching levels included in the quantiles argument

Slope_Trend

Estimated mean annual percent change over the trend time-period, according to the slope of a linear regression through the log-transformed annual indices

Slope_Trend_Q_quantiles

quantiles of the posterior distribution of Percent Change estimates, matching levels included in the quantiles argument

prob_decrease_X_percent

proportion of the posterior distribution of Percent_Change that is below the percentage values supplied in prob_decrease

prob_increase_X_percent

proportion of the posterior distribution of Percent_Change that is above the percentage values supplied in prob_increase

Relative_Abundance

Mean of the annual index values across all years. An estimate of the average relative abundance of the species in the region. Can be interepreted as the predicted average count of the species in an average year on an average route by an average observer, for the years, routes, and observers in the existing data

Observed_Relative_Abundance

Mean of the observed annual counts of birds across all routes and all years. An alternative estimate of the average relative abundance of the species in the region. For composite regions (i.e., anything other than stratum-level estimates) this average count is calculated as an area-weighted average across all strata included

Number_of_Strata

The number of strata included in the region

Width_of_X_percent_Credible_Interval

Width (in percent/year) of the credible interval on the Trend calculation. Calculated for the widest credible interval requested in quantiles argument. Default is 95 percent CI (i.e., Trend_Q0.975 - Trend_Q0.025)

Width_of_X_percent_Credible_Interval_Slope

Width (in percent/year) of the credible interval on the Trend calculation for the slope-based trend. Calculated for the widest credible interval requested in quantiles argument. Default is 95 percent CI (i.e., Slope_Trend_Q0.975 - Slope_Trend_Q0.025)

Number_of_Routes

The number of unique BBS routes included in the annual indices for this region and species, i.e., number of routes for this region and species for the years since generate_indices(startyear)

Mean_Number_of_Routes

The average number of BBS routes across years contributing data for this region and species

backcast_flag

approximate proportion of the included species range*years that are supported by data in a given region and year, e.g., 1.0 = data cover full time-series, 0.75 = data cover 75 percent of time-series. Only calculated if max_backcast != NULL

Details

The posterior probabilities can be calculated for a percent-decrease (prob_decrease) and/or percent-increase (prob_increase) if desired. These calculate the probability that the populationhas decreased/increased by at least the amount specified.

For example, a prob_increase = 100 would result in the calculation of the probability that the population has increased by more than 100% (i.e., doubled) over the period of the trend.

Alternatively, a prob_decrease = 50 would result in the calculation of the probability that the population has decreased by more than 50% (i.e., less than half of the population remains) over the period of the trend.

Examples

# Toy example with Pacific Wren sample data
# First, stratify the sample data
s <- stratify(by = "bbs_cws", sample_data = TRUE)
#> Using 'bbs_cws' (standard) stratification
#> Using sample BBS data...
#> Using species Pacific Wren (sample data)
#> Stratifying data...
#>   Combining BCR 7 and NS and PEI...
#>   Renaming routes...

# Prepare the stratified data for use in a JAGS model.
d <- prepare_data(s, min_year = 2009, max_year = 2018)

# Now run the model (fast but not good, just for illustration)
m <- run_model(d, model = "first_diff",
               iter_sampling = 5, iter_warmup = 5, chains = 2)
#> Running MCMC with 2 chains, at most 4 in parallel...
#> 
#> Chain 1 WARNING: No variance estimation is 
#> Chain 1          performed for num_warmup < 20 
#> Chain 1 Iteration: 1 / 10 [ 10%]  (Warmup) 
#> Chain 1 Iteration: 6 / 10 [ 60%]  (Sampling) 
#> Chain 1 Iteration: 10 / 10 [100%]  (Sampling) 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 2 WARNING: No variance estimation is 
#> Chain 2          performed for num_warmup < 20 
#> Chain 2 Iteration: 1 / 10 [ 10%]  (Warmup) 
#> Chain 2 Iteration: 6 / 10 [ 60%]  (Sampling) 
#> Chain 2 Iteration: 10 / 10 [100%]  (Sampling) 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpkaiFyy/model-462c337e1f9a.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 1 finished in 0.0 seconds.
#> Chain 2 finished in 0.0 seconds.
#> 
#> Both chains finished successfully.
#> Mean chain execution time: 0.0 seconds.
#> Total execution time: 0.2 seconds.
#> 
#> Warning: 10 of 10 (100.0%) transitions ended with a divergence.
#> See https://mc-stan.org/misc/warnings for details.
#> Saving model output to /home/runner/work/bbsBayes/bbsBayes/docs/reference/BBS_STAN_first_diff_hier_2022-11-05_01.rds

# Generate the continental and stratum indices
i <- generate_indices(model_output = m)
#> Processing region stratum
#> Processing region continent

# Now, generate the trends
t <- generate_trends(i)