Save model to file

Save a predefined Stan model file to a local text file for editing. These files can then be used in run_models() by specifying the model_file argument.

Usage

model_to_file(model, model_variant, dir, overwrite = FALSE)

Arguments

model

Character. Type of model to use, must be one of "first_diff" (First Differences), "gam" (General Additive Model), "gamye" (General Additive Model with Year Effect), or "slope" (Slope model).

model_variant

Character. Model variant to use, must be one of "nonhier" (Non-hierarchical), "hier" (Hierarchical), or "spatial" (Spatially explicit).

dir

Character. Directory where file should be saved.

overwrite

Logical. Whether to overwrite an existing copy of the model file.

Value

File path to copy of the model file.

Examples

# Save the Slope model in temp directory
model_to_file(model = "slope", model_variant = "spatial", dir = tempdir())
#> Copying model file slope_spatial_bbs_CV.stan to /tmp/Rtmp7u0qSW/slope_spatial_bbs_CV_COPY.stan
#> [1] "/tmp/Rtmp7u0qSW/slope_spatial_bbs_CV_COPY.stan"

# Overwrite an existing copy
model_to_file(model = "slope", model_variant = "spatial", dir = tempdir(),
              overwrite = TRUE)
#> Copying model file slope_spatial_bbs_CV.stan to /tmp/Rtmp7u0qSW/slope_spatial_bbs_CV_COPY.stan
#> [1] "/tmp/Rtmp7u0qSW/slope_spatial_bbs_CV_COPY.stan"

# Clean up
unlink(file.path(tempdir(), "slope_spatial_bbs_CV_COPY.stan"))