Run Bayesian model

Run Bayesian model

Usage

run_model(
  prepped_data,
  model,
  model_variant = "hier",
  model_file = NULL,
  spatial_data = NULL,
  heavy_tailed = TRUE,
  n_knots = NULL,
  basis = "mgcv",
  use_pois = FALSE,
  calculate_nu = FALSE,
  calculate_log_lik = FALSE,
  calculate_CV = FALSE,
  refresh = 100,
  chains = 4,
  parallel_chains = 4,
  iter_warmup = 1000,
  iter_sampling = 1000,
  adapt_delta = 0.95,
  max_treedepth = 14,
  init = NULL,
  init_only = FALSE,
  output_basename = NULL,
  output_dir = ".",
  save_output = TRUE,
  quiet = FALSE,
  jags_data,
  inits,
  parameters_to_save,
  track_n,
  n_adapt,
  n_burnin,
  n_thin,
  n_chains,
  n_saved_steps,
  n_iter,
  modules,
  parallel,
  n_cores,
  ...
)

Arguments

prepped_data

List. Output of prepare_data().

model

Character. Type of model to use, must be one of "first_diff" (First Differences), "gam" (General Additive Model), "gamye" (General Additive Model with Year Effect), or "slope" (Slope model).

model_variant

Character. Model variant to use, must be one of "nonhier" (Non-hierarchical), "hier" (Hierarchical), or "spatial" (Spatially explicit).

model_file

Character. Location of a custom Stan model file to use.

spatial_data

List. Output of prepare_spatial().

heavy_tailed

Logical. Whether extra-Poisson error distributions should be modelled as a t-distribution, with heavier tails than the standard normal distribution. Default is TRUE. Recent results suggest this is best even though it requires much longer convergence times. Can only be used with Negative Binomial models (i.e. use_pois = FALSE).

n_knots

Numeric. Number of knots for "gam" and "gamye" models

basis

Character. Basis function to use for GAM smooth, one of "original" or "mgcv". Default is "original", the same basis used in Smith and Edwards 2020. "mgcv" is an alternate that uses the "tp" basis from the package mgcv (also used in brms, and rstanarm). If using the "mgcv" option, the user may want to consider adjusting the prior distributions for the parameters and their precision.

use_pois

Logical. Whether to use an Over-Dispersed Poisson model (TRUE) or an Negative Binomial model (FALSE, default).

calculate_nu

Logical. Whether to calculate the nu parameter as a factor of gamma(2, 0.1). Default is FALSE.

calculate_log_lik

Logical. Whether to calculate point-wise log-likelihood of the data given the model. Default is FALSE.

calculate_CV

Logical. Whether to use cross validation.

refresh

Numeric. Passed to cmdstanr::sample(). Number of iterations between screen updates. If 0, only errors are shown.

chains

Numeric. Passed to cmdstanr::sample(). Number of Markov chains to run.

parallel_chains

Numeric. Passed to cmdstanr::sample(). Maximum number of chains to run in parallel.

iter_warmup

Numeric. Passed to cmdstanr::sample(). Number of warmup iterations per chain.

iter_sampling

Numeric. Passed to cmdstanr::sample(). Number of sampling (post-warmup) iterations per chain.

adapt_delta

Numeric. Passed to cmdstanr::sample(). The adaptation target acceptance statistic.

max_treedepth

Numeric. Passed to cmdstanr::sample(). The maximum allowed tree depth for the NUTS engine. See ?cmdstanr::sample.

init

(multiple options). An initialization method passed to the init argument of cmdstanr::sample(). See ?cmdstanr::sample for more details. The default (NULL) is for bbsBayes to supply initial parameter values according to model and model variant selected.

init_only

Logical. Whether to return the list of initial parameter values only (and not run the model). This is useful if you wish to see which init parameters are being supplied.

output_basename

Character. Name of the files created as part of the Stan model run and the final model output RDS file if save_output = TRUE.

output_dir

Character. Directory in which all model files will be created.

save_output

Logical. Whether or not to save the model output to file as an RDS object with all required data. Defaults to TRUE.

quiet

Logical. Suppress progress messages? Default FALSE

jags_data

Defunct.

inits

Defunct.

parameters_to_save

Defunct.

track_n

Defunct.

n_adapt

Defunct.

n_burnin

Defunct.

n_thin

Defunct.

n_chains

Defunct.

n_saved_steps

Defunct.

n_iter

Defunct.

modules

Defunct.

parallel

Defunct.

n_cores

Defunct.

...

Other arguments passed on to cmdstanr::sample()

Value

A list containing the model output (model_fit), meta data for the analysis (meta_data), meta data for the strata (meta_strata) and prepared data counts from prepare_data() (raw_data).

Details

There are two ways you can customize the model run. The first is to supply a custom model_file created with the copy_model_file() function and then edited by hand.

Second, you can supply a custom list of initialization parameters. You can supply these parameters in anyway that cmdstanr::sample() accepts the init argument. One way is to use the bbsBayes created parameters and modify them as needed. To do this, setup the run_model() function as you would normally use it, but add init_only = TRUE. This will return the list of initial definitions. You can edit this list and then run your model with run_model(... init = new_inits).

See the models article for more advanced examples and explanations.

Examples

s <- stratify(by = "bbs_cws", sample_data = TRUE)
#> Using 'bbs_cws' (standard) stratification
#> Using sample BBS data...
#> Using species Pacific Wren (sample data)
#> Stratifying data...
#>   Combining BCR 7 and NS and PEI...
#>   Renaming routes...
p <- prepare_data(s)

# Run model (quick and dirty)
m <- run_model(p, model = "first_diff", model_variant = "hier",
               iter_warmup = 20, iter_sampling = 20, chains = 2)
#> Running MCMC with 2 chains, at most 4 in parallel...
#> 
#> Chain 1 WARNING: There aren't enough warmup iterations to fit the 
#> Chain 1          three stages of adaptation as currently configured. 
#> Chain 1          Reducing each adaptation stage to 15%/75%/10% of 
#> Chain 1          the given number of warmup iterations: 
#> Chain 1            init_buffer = 3 
#> Chain 1            adapt_window = 15 
#> Chain 1            term_buffer = 2 
#> Chain 1 Iteration:  1 / 40 [  2%]  (Warmup) 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 1 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 1 Exception: neg_binomial_2_log_lpmf: Precision parameter is inf, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 1 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 1 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 1 
#> Chain 2 WARNING: There aren't enough warmup iterations to fit the 
#> Chain 2          three stages of adaptation as currently configured. 
#> Chain 2          Reducing each adaptation stage to 15%/75%/10% of 
#> Chain 2          the given number of warmup iterations: 
#> Chain 2            init_buffer = 3 
#> Chain 2            adapt_window = 15 
#> Chain 2            term_buffer = 2 
#> Chain 2 Iteration:  1 / 40 [  2%]  (Warmup) 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is 0, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Informational Message: The current Metropolis proposal is about to be rejected because of the following issue:
#> Chain 2 Exception: neg_binomial_2_log_lpmf: Precision parameter is inf, but must be positive finite! (in '/tmp/RtmpWV7JXg/model-252d66cad0ec.stan', line 227, column 3 to column 40)
#> Chain 2 If this warning occurs sporadically, such as for highly constrained variable types like covariance matrices, then the sampler is fine,
#> Chain 2 but if this warning occurs often then your model may be either severely ill-conditioned or misspecified.
#> Chain 2 
#> Chain 2 Iteration: 21 / 40 [ 52%]  (Sampling) 
#> Chain 1 Iteration: 21 / 40 [ 52%]  (Sampling) 
#> Chain 2 Iteration: 40 / 40 [100%]  (Sampling) 
#> Chain 2 finished in 163.3 seconds.
#> Chain 1 Iteration: 40 / 40 [100%]  (Sampling) 
#> Chain 1 finished in 189.5 seconds.
#> 
#> Both chains finished successfully.
#> Mean chain execution time: 176.4 seconds.
#> Total execution time: 189.6 seconds.
#> 
#> Saving model output to /home/runner/work/bbsBayes/bbsBayes/docs/reference/BBS_STAN_first_diff_hier_2022-12-16_01.rds

# Clean up (remove model files)
unlink(list.files(pattern = paste0("BBS_STAN_first_diff_hier_", Sys.Date())))